bytes in ASTs: use Cow<_, [u8]> instead of &[u8]
The motivation is to make it easily possible to produce "owned" AST
values that are not the result of parsing a single slice of input
bytes. This is useful for composing emails, but also fuzzing, etc.

We also use the bounded_static crate to derive a ToStatic trait, which
makes it easy to go from a "borrowed" version of an AST type Foo<'a>
to an "owned" copy Foo<'static>.